pyproject.toml
[tool.poetry]
name = "usertemplateapi"
version = "0.1.0"
description = ""
authors = ["Davi Lucciola <davilucciola14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
sqlmodel = "^0.0.11"
uvicorn = "^0.24.0.post1"
python-decouple = "^3.8"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
httpx = "^0.25.0"
email-validator = "^2.1.0.post1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
